+++ shared/toks-components/src/components/index.ts
@@ -4,6 +4,8 @@ export * from './BackButton';
 export * from './Text';
 export * from './SSRSuspense';
 
+export * from './Editor';
+
 export * from 'primereact/accordion';
 export * from 'primereact/autocomplete';
 export * from './UserAvatar';
